Authentication system working!
Basic authentication system is working with the database. Security level is very low, but the password is at least hashed. Authenticated user sessions are stored site wise and can be identified by calling the session user_name. Removed the game from the index.php to make it clearer how the system works before implementing it to the game. I will start working on security and error handling from now.